Bank-UI/Cua-ui/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
python-dotenv==1.0.0
playwright==1.40.0
pillow==10.1.0
